:building_construction: Move new component to another folder
