+++ src/components/card.tsx
@@ -1,25 +1,29 @@
-import { Link } from 'gatsby'
-import React from 'react'
-
-import styled, { createGlobalStyle } from 'styled-components'
+import React, { ReactNode } from 'react'
 
 import Img, { FluidObject } from 'gatsby-image'
 
-interface PropsInterface {
-  author?: {
-    user: string
-    name: string
-    facebook: string
-    twitter?: string
+import styled, { createGlobalStyle } from 'styled-components'
+
+import { Box, Card, Heading, Link, Text } from 'rebass'
+
+interface IProps {
+  children?: ReactNode
+  blog: {
+    title: string
+    subtitle?: string
+    date?: string
+    banner: {
+      childImageSharp: {
+        fluid: FluidObject
+      }
+    }
   }
-  banner: FluidObject
-  date?: string
-  featured?: boolean
-  link: boolean
-  slug: string
-  subtitle?: string
-  title: string
-  content?: string
+  author?: {
+    name: string
+    user: string
+  }
+  slug?: string
+  type: string
 }
 
 const GlobalStyle = createGlobalStyle`
@@ -41,7 +45,7 @@ const GlobalStyle = createGlobalStyle`
 
   p,
   ol,
-  table {
+  table>tbody>tr>td {
     color: rgba(0, 0, 0, 0.6)
   }
 
@@ -67,143 +71,35 @@ const GlobalStyle = createGlobalStyle`
   }
 `
 
-const Container = styled.div`
-  background: #fff;
-  margin-top: 25px;
-  padding: 40px;
-
-  @media only screen and (min-width: 960px) {
-    margin-top: 45px;
-    padding: 70px;
-  }
-`
-
-const Teaser = styled.div`
-  margin: -40px -40px -40px -40px;
-  display: block;
-
-  @media only screen and (min-width: 960px) {
-    margin: -70px -70px -70px -70px;
-  }
-`
-
-const Slug = styled.div`
-  background-color: #333;
-  color: #fff;
-  font-weight: bold;
-  letter-spacing: .04em;
-  padding: .25em .5em;
-  position: relative;
-  text-transform: uppercase;
-  z-index: 2;
-  font-size: 26px;
-  margin-right: 10px;
-  display: inline;
-
-  @media only screen and (max-width: 540px) {
-    font-size: 14px;
-  }
-`
-
 const Banner = styled(Img)`
-  margin-top: -45.5px;
-  z-index: 1;
-
-  @media only screen and (max-width: 540px) {
-    margin-top: -25.5px;
-  }
+  border-radius: 6px 6px 0 0;
 `
 
-const Header = styled.div`
-  margin-top: 80px;
+const Component: React.SFC<IProps> = props => {
+  const {author, blog, children, slug, type} = props
+  const {title, subtitle, banner, date} = blog
 
-  @media only screen and (min-width: 960px) {
-    margin-top: 110px;
-  }
-`
-
-const Title = styled.div`
-  font-size: 36px;
-  line-height: 46px;
-  font-weight: 400;
-  text-transform: none;
-  
-  &>a {
-    color: rgba(0, 0, 0, 1.0);
-    text-decoration: none;
-    cursor: pointer;
-  }
-
-  @media only screen and (min-width: 960px) {
-    font-size: 38px;
-  }
-`
-
-const Subtitle = styled.div`
-  color: rgba(0,0,0,.6);
-  font-style: normal;
-  font-weight: 400;
-`
-
-const Meta = styled.div`
-  color: rgba(0, 0, 0, 0.6);
-  font-size: 16px;
-  line-height: 20px;
-  margin-bottom: 25px;
-  margin-top: 25px;
-
-  &>a {
-    color: rgba(0, 0, 0, 0.8);
-  }
-`
-
-const Content = styled.div`
-  &>ul>li {
-    color: #6f6f6f;
-  }
-`
-
-const Card: React.SFC<PropsInterface> = props => {
-  const {author, banner, date, featured, link, slug, subtitle, title, content, children} = props
-
-  const processedBanner =
-    link === true ? (
-      <Link to={slug} aria-label={`card-banner-link-${slug}`}>
-        <Banner fluid={banner} />
-      </Link>
-    ) : (
-      <Banner fluid={banner} />
-    )
-
-  const processedTitle =
-    link === true ? (
-      <Link to={slug} aria-label={`card-title-link-${slug}`}>
-        {title}
-      </Link>
-    ) : (
-      title
-    )
+  const cardBanner = <Banner fluid={banner.childImageSharp.fluid} />
+  const cardTitle = <Heading fontSize={type === 'listing' ? [24, 26, 28, 30] : type === 'post' ? [30, 32, 34, 36] : 38} fontWeight={400} color={`rgb(0, 0, 0)`}>{title}</Heading>
 
   return (
-    <Container key={slug}>
+    <Card backgroundColor={`rgb(255, 255, 255)`} borderRadius={6} boxShadow={`0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)`}>
       <GlobalStyle />
-      <Teaser>
-        {featured && <Slug>featured</Slug>}
-        {processedBanner}
-      </Teaser>
-      <Header>
-        <Title>{processedTitle}</Title>
-        {date && author && (
-          <Meta>
-            Written by <Link to={'/author/' + author.user}>{author.name}</Link> on {date}
-          </Meta>
-        )}
-        {subtitle && <Subtitle>{subtitle}</Subtitle>}
-      </Header>
-      {content && <Content dangerouslySetInnerHTML={{__html: content}} />}
+      <Box>
+        {slug ? <Link href={slug}>{cardBanner}</Link> : cardBanner}
+      </Box>
+      <Box p={4}>
+        {slug ? <Link href={slug}>{cardTitle}</Link> : cardTitle}
+        {date && author ? (
+          <Text fontSize={[14, 16]} mt={3} color={`rgba(0, 0, 0, 0.6)`}>
+            Written by <Link href={'/author/' + author.user} color={`rgba(0, 0, 0, 0.8)`}>{author.name}</Link> on {date}
+          </Text>
+        ) : null}
+        {subtitle ? <Text fontSize={[16, 17]} mt={3} color={`rgba(0, 0, 0, 0.6)`}>{subtitle}</Text> : null}
+      </Box>
       {children}
-    </Container>
+    </Card>
   )
 }
 
-export { Card }
+export default Component
