+++ src/templates/category-blog.tsx
@@ -9,9 +9,9 @@ import { FluidObject } from 'gatsby-image'
 
 import { Box, Flex } from 'rebass'
 
-import { Chip } from '../components/chip'
-import Card from '../components/new/card'
-import Pagination from '../components/new/pagination'
+import Card from '../components/card'
+import Chip from '../components/chip'
+import Pagination from '../components/pagination'
 
 interface IAuthor {
   node: {
