+++ src/templates/category-list.tsx
@@ -5,8 +5,8 @@ import { FluidObject } from 'gatsby-image'
 
 import { Box, Flex } from 'rebass'
 
-import { Chip } from '../components/chip'
-import Card from '../components/new/card'
+import Card from '../components/card'
+import Chip from '../components/chip'
 
 interface PropsInterface {
   location: object
